Emit event when last_activity_at is updated.
TNL-3303
