+++ lms/djangoapps/teams/tests/test_views.py
@@ -418,7 +418,7 @@ class TestListTeamsAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the team listing API endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestListTeamsAPI, self).setUp('teams.views.tracker')
+        super(TestListTeamsAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
@@ -592,7 +592,7 @@ class TestCreateTeamAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the team creation endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestCreateTeamAPI, self).setUp('teams.views.tracker')
+        super(TestCreateTeamAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
@@ -803,7 +803,7 @@ class TestDeleteTeamAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the team delete endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestDeleteTeamAPI, self).setUp('teams.views.tracker')
+        super(TestDeleteTeamAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
@@ -853,7 +853,7 @@ class TestUpdateTeamAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the team update endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestUpdateTeamAPI, self).setUp('teams.views.tracker')
+        super(TestUpdateTeamAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
@@ -1182,7 +1182,7 @@ class TestCreateMembershipAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the membership creation endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestCreateMembershipAPI, self).setUp('teams.views.tracker')
+        super(TestCreateMembershipAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
@@ -1346,7 +1346,7 @@ class TestDeleteMembershipAPI(EventTestMixin, TeamAPITestCase):
     """Test cases for the membership deletion endpoint."""
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(TestDeleteMembershipAPI, self).setUp('teams.views.tracker')
+        super(TestDeleteMembershipAPI, self).setUp('teams.utils.tracker')
 
     @ddt.data(
         (None, 401),
